added session history

+++ db/migrations/20201005201319-create-review.js
@@ -10,7 +10,7 @@ module.exports = {
       },
       rating: {
         allowNull: false,
-        type: Sequelize.INTEGER(5)
+        type: Sequelize.INTEGER
       },
       description: {
         allowNull: false,
